Code file structure refactor + state management and UI controller
