+++ requirements.txt
@@ -1,0 +1,25 @@
+contourpy==1.1.1
+CTkMessagebox==2.5
+customtkinter==5.2.2
+cycler==0.12.1
+darkdetect==0.8.0
+dynamixel-sdk==3.7.31
+EasyPySpin==2.0.1
+fonttools==4.49.0
+importlib_resources==6.1.2
+kiwisolver==1.4.5
+matplotlib==3.7.5
+numpy==1.24.4
+opencv-python==4.9.0.80
+packaging==23.2
+pillow==10.2.0
+psutil==5.9.8
+pyparsing==3.1.1
+pyserial==3.5
+python-dateutil==2.9.0.post0
+qtm==2.1.0
+six==1.16.0
+# spinnaker-python @ file:///src/spinnaker_python-3.1.0.79-cp38-cp38-macosx_10_14_x86_64.whl#sha256=00ab9b8bee52c3fd8a4535ff3c2c97282c31c6ab84bae5df9272cf678d3f5b55
+zipp==3.17.0
+
+
