+++ src/ui/views/theia_control_window.py
